+++ base/build.gradle
@@ -6,7 +6,8 @@ dependencies {
     kapt libs.dagger_android_compiler
 
     implementation libs.glide
-    implementation libs.threeten
+    // Marked as API as this is used by the ':localization' and ':db' modules
+    api libs.threeten
     // This is now marked as API because other modules need this to be that version
     api libs.design_support
