+++ details/src/main/kotlin/edu/artic/events/EventDetailViewModel.kt
@@ -5,7 +5,7 @@ import com.fuzz.rx.disposedBy
 import edu.artic.analytics.AnalyticsAction
 import edu.artic.analytics.AnalyticsTracker
 import edu.artic.analytics.ScreenCategoryName
-import edu.artic.base.utils.DateTimeHelper.Purpose.*
+import edu.artic.localization.util.DateTimeHelper.Purpose.*
 import edu.artic.db.models.ArticEvent
 import edu.artic.localization.LanguageSelector
 import edu.artic.viewmodel.NavViewViewModel
